Rerun Sumyyah approach 4

+++ src/torch_dti_utils.py
@@ -683,7 +683,7 @@ def quick_train(config, model, device, train_loader, optimizer, epoch, neg_to_po
 
         print('y.size()', y.size())
 
-        help_mask = np.around(np.array(y.to('cpu')) * train_mask)
+        help_mask = np.around(np.array(y.to('cpu')) * train_mask).astype(np.int)
 
         print('help_mask', help_mask.dtype)
